CAN-16: add check to justfile

+++ justfile
@@ -137,11 +137,17 @@ generate TARGET_DIR FLAG="":
 
 # Generate an SDK from a swagger.json and copy the output to the TARGET_DIR
 generate-cicd TARGET_DIR FLAG="":
+    set -ex
     mkdir -p {{TARGET_DIR}}
     mkdir -p ./generate/.output
     envsubst < generate/config-template.json > generate/.config.json
     cp ./generate/.openapi-generator-ignore ./generate/.output/.openapi-generator-ignore
-    cp ./generate/templates/description.{{APPLICATION_NAME}}.mustache ./generate/templates/description.mustache
+    # Basic check if file exists
+    if [ -f "./generate/templates/description.{{APPLICATION_NAME}}.mustache" ]; then
+        cp ./generate/templates/description.{{APPLICATION_NAME}}.mustache ./generate/templates/description.mustache
+    else
+        echo "No description template for {{ APPLICATION_NAME }} ... skipping."
+    fi
 
     ./generate/generate.sh ./generate ./generate/.output {{swagger_path}} .config.json
     rm -f generate/.output/.openapi-generator-ignore
